setup.py will discover version from git if needed
This might seem silly at first glance--for most things a default version
of 0.0.0 is fine. However, given the tightly-coupled nature of
cirrus-geo and cirrus-lib, we often want to install cirrus-lib to the
same dev venv as cirrus-geo, which causes a conflict as cirrus-lib 0.0.0
is below the version specified as a dependency of cirrus-geo. As a
result, any reinstallation of cirrus-geo to the venv will cause the
local dev version of cirrus-lib to be uninstalled.

If instead we find the most recent tag of cirrus-lib and base our
version on that, then we do not have an issue--local development changes
will simply be appended to the last tagged version and pip will be
happy.